chars aren't guaranteed to be signed, quell mipspro warning and hope it isn't faulty logic
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@12574 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/DynamicColor.cxx
@@ -201,7 +201,7 @@ void DynamicColor::finalize()
   for (int c = 0; c < 4; c++) {
     sequenceParams& seq = channels[c].sequence;
     for (unsigned int i = 0; i < seq.count; i++) {
-      if (seq.list[i] < colorMin) {
+      if ((int)seq.list[i] < colorMin) {
 	seq.list[i] = colorMin;
       } else if (seq.list[i] > colorMax) {
 	seq.list[i] = colorMax;
